pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.30",
    "langchain-gemini>=0.1.1",
    "langchain-anthropic>=0.3.30",
    "langchain-openrouter>=0.1.0",
    "python-dotenv>=1.1.1",
    "rich>=14.3.3",
    "langchain-tavily>=0.2.13",
    "tavily-python>=0.7.12",
    "rich>=14.3.3",
    "langchain-community>=0.4.1",
    "langchain-pinecone>=0.2.13",
]
